@@ -1,15 +1,19 @@
 package com.ead.payment.configs.security;
 
+import jakarta.servlet.DispatcherType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
 import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
 import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -17,12 +21,15 @@ import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 @Configuration
-@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableMethodSecurity
 @EnableWebSecurity
 public class WebSecurityConfig {
 
 	@Autowired
 	AuthenticationEntryPointImpl authenticationEntryPoint;
+
+	@Autowired
+	AccesDeniedHandlerImpl accesDeniedHandlerImpl;
 
 	@Bean
 	public AuthenticationJwtFilter authenticationJwtFilter() {
@@ -39,17 +46,28 @@ public class WebSecurityConfig {
 	}
 
 	@Bean
+	public DefaultMethodSecurityExpressionHandler expressionHandler() {
+		DefaultMethodSecurityExpressionHandler expressionHandler = new DefaultMethodSecurityExpressionHandler();
+		expressionHandler.setRoleHierarchy(roleHierarchy());
+		return expressionHandler;
+	}
+
+	@Bean
 	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 		http
-				.exceptionHandling().authenticationEntryPoint(authenticationEntryPoint)
-				.and()
-				.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-				.and()
-				.authorizeRequests()
-				.anyRequest().authenticated()
-				.and()
-				.csrf().disable();
+				.csrf(AbstractHttpConfigurer::disable)
+				.exceptionHandling((exceptionHandling) ->
+						exceptionHandling
+								.accessDeniedHandler(accesDeniedHandlerImpl)
+								.authenticationEntryPoint(authenticationEntryPoint))
+				.sessionManagement(session ->
+						session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
+				.authorizeHttpRequests((authorize) -> authorize
+						.dispatcherTypeMatchers(DispatcherType.ERROR).permitAll()
+						.anyRequest().authenticated()
+				);
 		http.addFilterBefore(authenticationJwtFilter(), UsernamePasswordAuthenticationFilter.class);
+
 		return http.build();
 	}
 	@Bean
