@@ -2,7 +2,7 @@ package com.ead.payment.dtos;
 
 import org.hibernate.validator.constraints.br.CPF;
 
-import javax.validation.constraints.*;
+import jakarta.validation.constraints.*;
 import java.math.BigDecimal;
 
 public class PaymentRequestDto {
